Fixed up some code.
Got printing working. That was a pain. And scary. I decided it was best
if that function is unsafe.

Also realized I needed to _borrow_ the ParserInputPosition, since it
contained a pointer. That was a fun segfault to try and figure out.

Used mopa to have my own Any, so I don't need to use _unbox_ or some
other hack like that. Still need to use to_sub() until I can figure out
a more clever way to accomplish that.

Made a giant mess with macros. It is technically very clean, but I'm not
a fan of the fact that I can't format the code automatically.

Made the error type non_exhaustive, so we can add stuff without breaking
code.

Shamelessly stole some tests from the libcypher-parser C library. I'm
not done doing that, either, so I will most certainly be adding more
from that as a sanity check. I plan to do that to ensure correct
behavior.